+++ sdk/resourcemanager/appservice/armappservice/models.go
@@ -477,10 +477,12 @@ type ApplicationStackResource struct {
 }
 
 type ArcConfiguration struct {
-	ArtifactStorageAccessMode    *string
-	ArtifactStorageClassName     *string
-	ArtifactStorageMountPath     *string
-	ArtifactStorageNodeName      *string
+	ArtifactStorageAccessMode *string
+	ArtifactStorageClassName  *string
+	ArtifactStorageMountPath  *string
+	ArtifactStorageNodeName   *string
+
+	// Property to select Azure Storage type. Available options: blobContainer.
 	ArtifactsStorageType         *StorageType
 	FrontEndServiceConfiguration *FrontEndConfiguration
 	KubeConfig                   *string
@@ -3434,6 +3436,18 @@ type FrontEndConfiguration struct {
 	Kind *FrontEndServiceType
 }
 
+// FunctionAppConfig - Function app configuration.
+type FunctionAppConfig struct {
+	// Function app deployment configuration.
+	Deployment *FunctionsDeployment
+
+	// Function app runtime settings.
+	Runtime *FunctionsRuntime
+
+	// Function app scale and concurrency settings.
+	ScaleAndConcurrency *FunctionsScaleAndConcurrency
+}
+
 // FunctionAppMajorVersion - Function App stack major version.
 type FunctionAppMajorVersion struct {
 	// READ-ONLY; Function App stack major version (display only).
@@ -3633,6 +3647,87 @@ type FunctionSecrets struct {
 
 	// Trigger URL.
 	TriggerURL *string
+}
+
+// FunctionsAlwaysReadyConfig - Sets the number of 'Always Ready' instances for a function group or a specific function.
+type FunctionsAlwaysReadyConfig struct {
+	// Sets the number of 'Always Ready' instances for a given function group or a specific function. For additional information
+	// see https://aka.ms/flexconsumption/alwaysready.
+	InstanceCount *float32
+
+	// Either a function group or a function name is required. For additional information see https://aka.ms/flexconsumption/alwaysready.
+	Name *string
+}
+
+// FunctionsDeployment - Configuration section for the function app deployment.
+type FunctionsDeployment struct {
+	// Storage for deployed package used by the function app.
+	Storage *FunctionsDeploymentStorage
+}
+
+// FunctionsDeploymentStorage - Storage for deployed package used by the function app.
+type FunctionsDeploymentStorage struct {
+	// Authentication method to access the storage account for deployment.
+	Authentication *FunctionsDeploymentStorageAuthentication
+
+	// Property to select Azure Storage type. Available options: blobContainer.
+	Type *StorageType
+
+	// Property to set the URL for the selected Azure Storage type. Example: For blobContainer, the value could be https://.blob.core.windows.net/.
+	Value *string
+}
+
+// FunctionsDeploymentStorageAuthentication - Authentication method to access the storage account for deployment.
+type FunctionsDeploymentStorageAuthentication struct {
+	// Use this property for StorageAccountConnectionString. Set the name of the app setting that has the storage account connection
+	// string. Do not set a value for this property when using other
+	// authentication type.
+	StorageAccountConnectionStringName *string
+
+	// Property to select authentication type to access the selected storage account. Available options: SystemAssignedIdentity,
+	// UserAssignedIdentity, StorageAccountConnectionString.
+	Type *AuthenticationType
+
+	// Use this property for UserAssignedIdentity. Set the resource ID of the identity. Do not set a value for this property when
+	// using other authentication type.
+	UserAssignedIdentityResourceID *string
+}
+
+// FunctionsRuntime - Function app runtime name and version.
+type FunctionsRuntime struct {
+	// Function app runtime name. Available options: dotnet-isolated, node, java, powershell, python, custom
+	Name *RuntimeName
+
+	// Function app runtime version. Example: 8 (for dotnet-isolated)
+	Version *string
+}
+
+// FunctionsScaleAndConcurrency - Scale and concurrency settings for the function app.
+type FunctionsScaleAndConcurrency struct {
+	// 'Always Ready' configuration for the function app.
+	AlwaysReady []*FunctionsAlwaysReadyConfig
+
+	// Set the amount of memory allocated to each instance of the function app in MB. CPU and network bandwidth are allocated
+	// proportionally.
+	InstanceMemoryMB *float32
+
+	// The maximum number of instances for the function app.
+	MaximumInstanceCount *float32
+
+	// Scale and concurrency settings for the function app triggers.
+	Triggers *FunctionsScaleAndConcurrencyTriggers
+}
+
+// FunctionsScaleAndConcurrencyTriggers - Scale and concurrency settings for the function app triggers.
+type FunctionsScaleAndConcurrencyTriggers struct {
+	// Scale and concurrency settings for the HTTP trigger.
+	HTTP *FunctionsScaleAndConcurrencyTriggersHTTP
+}
+
+// FunctionsScaleAndConcurrencyTriggersHTTP - Scale and concurrency settings for the HTTP trigger.
+type FunctionsScaleAndConcurrencyTriggersHTTP struct {
+	// The maximum number of concurrent HTTP trigger invocations per instance.
+	PerInstanceConcurrency *float32
 }
 
 // GeoRegion - Geographical region.
@@ -7916,6 +8011,9 @@ type SiteProperties struct {
 	// true if the app is enabled; otherwise, false. Setting this value to false disables the app (takes the app offline).
 	Enabled *bool
 
+	// Configuration specific of the Azure Function app.
+	FunctionAppConfig *FunctionAppConfig
+
 	// HttpsOnly: configures a web site to accept only https requests. Issues redirect for http requests
 	HTTPSOnly *bool
 
@@ -7971,6 +8069,9 @@ type SiteProperties struct {
 	// the form
 	// /subscriptions/{subscriptionName}/resourceGroups/{resourceGroupName}/providers/Microsoft.Network/virtualNetworks/{vnetName}/subnets/{subnetName}
 	VirtualNetworkSubnetID *string
+
+	// To enable Backup and Restore operations over virtual network
+	VnetBackupRestoreEnabled *bool
 
 	// To enable accessing content over virtual network
 	VnetContentShareEnabled *bool
@@ -9343,14 +9444,40 @@ type SupportTopic struct {
 	PesID *string
 }
 
-// SwiftVirtualNetwork - Swift Virtual Network Contract. This is used to enable the new Swift way of doing virtual network
-// integration.
+// SwiftVirtualNetwork - Swift Virtual Network Contract
 type SwiftVirtualNetwork struct {
+	// Resources (sites and serverfarms) allocated to this subnet
+	ResourceAllocation *SwiftVirtualNetworkResourceAllocation
+
+	// IP allocation for a Swift Subnet
+	SubnetIPAllocation *SwiftVirtualNetworkSubnetIPAllocation
+
+	// The Virtual Network subnet's Azure resource ID
+	SubnetResourceID *string
+
+	// Boolean flag indicating whether Swift integration is supported
+	SwiftSupported *bool
+
+	// Swift Connection allocation for an App Service Plan
+	VnetConnectionAllocation *SwiftVirtualNetworkVnetConnectionAllocation
+}
+
+// SwiftVirtualNetworkCollection - Collection of Swift Virtual Networks
+type SwiftVirtualNetworkCollection struct {
+	// REQUIRED; Collection of resources.
+	Value []*SwiftVirtualNetwork
+
+	// READ-ONLY; Link to next page of resources.
+	NextLink *string
+}
+
+// SwiftVirtualNetworkProxy - Swift Virtual Network Contract (Proxy resource)
+type SwiftVirtualNetworkProxy struct {
 	// Kind of resource.
 	Kind *string
 
-	// SwiftVirtualNetwork resource specific properties
-	Properties *SwiftVirtualNetworkProperties
+	// Swift Virtual Network Contract
+	Properties *SwiftVirtualNetwork
 
 	// READ-ONLY; Resource Id.
 	ID *string
@@ -9362,14 +9489,31 @@ type SwiftVirtualNetwork struct {
 	Type *string
 }
 
-// SwiftVirtualNetworkProperties - SwiftVirtualNetwork resource specific properties
-type SwiftVirtualNetworkProperties struct {
-	// The Virtual Network subnet's resource ID. This is the subnet that this Web App will join. This subnet must have a delegation
-	// to Microsoft.Web/serverFarms defined first.
-	SubnetResourceID *string
+// SwiftVirtualNetworkResourceAllocation - Resources (sites and serverfarms) allocated to this subnet
+type SwiftVirtualNetworkResourceAllocation struct {
+	// The list of all ServerFarms' Azure resource ID that are using the VNET
+	ConnectedServerFarmsID []*string
 
-	// A flag that specifies if the scale unit this Web App is on supports Swift integration.
-	SwiftSupported *bool
+	// The list of all Sites' Azure resource ID that are using the VNET
+	ConnectedSitesID []*string
+}
+
+// SwiftVirtualNetworkSubnetIPAllocation - IP allocation for a Swift Subnet
+type SwiftVirtualNetworkSubnetIPAllocation struct {
+	// The number of IP addresses available in the Swift subnet address space
+	SubnetIPAddressesAvailable *int32
+
+	// The number of IP addresses that have already been used in the Swift subnet address space
+	SubnetIPAddressesUsed *int32
+}
+
+// SwiftVirtualNetworkVnetConnectionAllocation - Swift Connection allocation for an App Service Plan
+type SwiftVirtualNetworkVnetConnectionAllocation struct {
+	// The maximum number of Swift Connections available for an App Service Plan
+	VnetConnectionsMax *int32
+
+	// The number of Swift Connections used for an App Service Plan
+	VnetConnectionsUsed *int32
 }
 
 // Template - Container App versioned application definition. Defines the desired state of an immutable revision. Any changes
@@ -9854,6 +9998,12 @@ type VirtualIPMapping struct {
 
 	// Virtual IP address.
 	VirtualIP *string
+}
+
+// VirtualNetworkIntegrationRequest - Virtual Network Integration request content.
+type VirtualNetworkIntegrationRequest struct {
+	// REQUIRED; The Azure Resource ID of the subnet
+	SubnetResourceID *string
 }
 
 // VirtualNetworkProfile - Specification for using a Virtual Network.